perf: init --help text to 350ms
